amcl: laser: add off-map and non-free factors
Add a factor which is used for any particle which is off the map (out of
the map bounds entirely).

Add a separate factor to be used when a particle is on a non-free-space
of the map.

Add a radius which can be set to interpolate between non-free-space
factor and 1.0 based on the distance the pose is from the nearest
non-free space (up the set radius).

Abstract applying the model from updating the filter to wrap the model
applications so the code to apply the map factors is in one place.

Default to factors of 1.0, equivalent to this code being deactivated.