costmap_3d: delete out-of-bounds after updateCosts
Delete the out-of-bounds information from the master costmap
after the calls to updateCosts. This way, it is not possible
for a buggy layer to leave out-of-bounds information in the
map by not honoring the bounds map passed into updateCosts.

Restructuring to deleting after updateCosts also has the
advantage of making the calculation of the delta bounds simpler.