costmap_2d: resetBoundingBox: wait until update
Have resetBoundingBox wait up to three times for the costmap
to be current after an update cycle. This way, planning can not
happen on empty or mostly empty costmaps. This is important
for not creeping up on obstacles in a local planner and not
creating straight-line global plans that ignore the static
map, due to it not being in the global costmap after a recovery
initiated call to resetBoundingBox.