costmap_3d: octomap plugin: fix bug when first map lost
If the first full map was lost, the plugin would wait forever
for the first full map. Fix this by ignoring the first normal
update (non-full map) on a subscription, but re-subscribing on
any subsequent update (indicating the first full map was lost).

The first full map may often be lost when starting up a large
single-threaded spinning node (such as move_base).