costmap_3d: have query object handle interior collisions
FCL does not handle octomap cells inside the mesh, it treats the mesh
as a surface. For costmap purposes, it is much better to treat the
mesh as a solid. Emulate this behavior by detecting results inside
the volume using PCL's crop hull filter.

For collisions and distance, this gives correct results. For signed
distance, the results are only somewhat correct, in that the deepest
penetration is not found, just some penetration depth.