costmap_3d: add third "milli" cache in queries
Add a third in-between cache called the "milli" cache.
This cache is only populated when a resulting distance is over a
threshold. It is another fast-path cache like the micro cache, if
the milli cache is hit, the box/triangle pair's distance is directly
used as the distance. The milli-cache binning parameters can be
looser than the micro-cache as the error is limited by the
milli-distance threshold.

This greatly improves performance of planners that generally keep the
queried poses away from objects by allowing a coarser-grained
fast-path cache for such cases