costmap_3d: mark costmap stale if base off map
Mark the costmap stale if the robot base is off the map.

This prevent autonomuous motion from planners that check the costmap
state when the robot base is off the costmap. If the base is off the
costmap, it will be travelling in places the costmap can not
represent, which is totally unsafe.