costmap_3d: apply micro cache when past threshold
Only apply the micro cache when the cached distance is beyond a
threshold.

Set the cache thresholds at object creation time to be 2.5 times the
spatial resolution of the caches.

Check the milli-cache first, as it is the most common cache to hit on
the fast path. Then check the micro-cache and update the upper bound
pose distance if it is smaller than the milli-cache result.