costmap_3d: add IfUnknown update type
Add a new Costmap3D update type, IfUnknown, which only updates
a layer in places where all the layers above have no data.
This is useful for adding static data from a static map but
allowing layers above to override any data in the static layer,
such as allowing layers above to override an occupied space
in the static map with a free space.