costmap_3d: add ray query action server and service
Add an action server and a service to query a ray at a pose, modeled
as a rectangular prism with infinte length starting at the given
pose. The ray query always uses a point for the "robot" mesh and
returns the distance to the nearest costmap voxel inside or on the
region queried.

To implement the ray query, the region of interest in the
Costmap3DQuery object had to be given a scale. The scale is not
exposed via the normal query action server and service as the
intended mechanism to issue ray queries is through the new
action server/service.

Also, a bug in the interatction between the caches and the region of
interest inside Costmap3DQuery had to be fixed. The bug was discovered
after implementation. Rather than back-port a fix prior to adding
the ray query, the bug fix is included in the new ray query commit.
The fix is to check that the costmap voxel in the cache entry is still
inside the region of interest at the new pose. If it is not, the cache
entry can not be used for the query.