costmap_3d: Allow alt to PCL+VTK for loading STL file
The normal way to load STL file from disk is to
 #include <pcl/io/vtk_lib_io.h> & call pcl::io::loadPolygonFileSTL().
But in order for PCL to provide that, it must depend on VTK, which is a
huge, sprawling library with tons of dependencies. That is both not
desirable and difficult to provide in a Yocto build.

So, we have published a tiny alternative implementation of
pcl::io::loadPolygonFileSTL() as a separate package called stl_loader.
This alternative can be selecting by setting -DUSE_STL_LOADER=ON as a
cmake option.