costmap_3d: query: only track stats when debug enabled
Only pay to track the statistics when debug is enabled. Tracking the
statistics causes unnecessary synchronization operations. Only pay for
these if the log will be printed.

Also, while using std::atomic is faster than using locking to
synchronize the counters for statistics tracking, all the statistics
variables are next to one another leading to false sharing. This could
be remedied by spreading them out to the cache line size, but since the
tracking is normally disabled do not bother with this optimization.