costmap_3d: query: speed up parallel buffered queries
Speed up parallel buffered queries by not holding the read lock on the
query object. Buffered queries only ever change when updateCostmap or
updateMeshResource are called, and it is up to the caller to ensure
there are no outstanding queries on the buffered query.

For this to work, it is necessary to create the non-lethal octree inside
checkCostmap instead of deferring it to the first non-lethal query. In
practice this is not a big deal as it does not cost much to construct
the non-lethal query and is better to do that at costmap update time
than deferring it which requires more synchronization.