costmap_3d: octree solver: readability, performance
Improve readability and performance by splitting the recursion into
phases. The first phase descends only the octree (same as before).
When the child of the octree is met, then the second phase is started
which checks the interior collision LUT. If this results in no
collision, then the mesh is descended. This improves readability by
making each phase of the search more obvious instead of as a bunch of
if's in one recursive function. It improves performance marginally by
skipping the ROI check while descending the mesh, and only performing
the interior collision check during the start of the mesh phase
avoiding a conditional.