costmap_3d: query: switch from FCL octree to octomap
No longer use customizations to FCL or FCL octree, instead directly use
octomap's octree. The FCL octree was being used to be able to store both
a free threshold and a lethal threshold (which is not an idea in
octomap's octree). This is not necessary for queries, however, as the
non-lethal queries already copy only the non-lethal parts of the tree
for more efficient querying. Instead, just set the occupied threshold
for the non-lethal tree to FREE so that the non-lethal nodes are counted
as occupied during the query.

By not using FCL for octree, there is no longer any pressure to maintain
the FCL API, such as using FCL's DistanceRequest or DistanceResult.
Instead, add a new DistanceRequest and DistanceResult inside of
Costmap3DQuery for the request/result structure information.

By not using the FCL API, the octomap box, box center and triangle mesh
id can just be directly saved inside the DistanceResult avoiding using
dynamic memory and shared pointers to return the box/triangle
information. This improves query performance. Also, by not using the FCL
octree, the octomap octree methods are directly called, which are
inline, so the hot-paths through the octree recursion are sped up
dramatically, further increasing performance.

By only saving the box center in the octomap's frame instead of a full
transform, it is more obvious that the octomap is not transformed by the
query, but left in the world frame and the two TF's are concatenated.

Part of this commit is to fix the spelling of primitive (it was
misspelled in the custom FCL modifications that are no longer being
used).