costmap_3d: fix TLS last cache on empty
If a query was run on an empty map (or one which is effectively empty),
a fast return path was taken. This path erroneously did not clear the
TLS last cache, which is necessary to prevent from reusing an invalid
result. This issue was caught when running with the other caches
disabled, which helped expose it, and has been present since the TLS
last cache was first implemented.