costmap_3d: query: fix cache erase corner cases
Fix a couple corner cases when erasing entries from the cache and milli
cache on costmap updates. The first issue was that the occupancy of the
found octomap cell was not being checked, so if the cell went from
occupied to free, but was still present in the map, the cache entry was
erroneously not erased. This could cause issues with navigation if the
attached octomap is tracking free space where an obstacle might appear
latched that is no longer on the costmap. As a part of the fix, use the
appropriate octree if the cache entry is from a non-lethal query. The
second issue was that the depth of the found node was not being
considered. Now that the octomap API search method can return the depth
of the found node, check that the found node's depth matches that in
the cache. If it does not, erase the entry.