WIP of binPose2: binning in 4-d by angles
Note: still need to cleanup and quantify how big each bin is. This could
be done by either figuring out analytically the sixteen hyper-cube
corners for each bin (Seems hard, esp since some of them are impossible,
but if at least one corner is possible, some poses bin). I think what
would be easier is to create a uniformly distributed structure on the
2-sphere (normal sphere) and use that as the rotation axes coupled with
uniform sampling on the rotation angle, then calculate the bin of each.
Then stats on each bin can be kept in a sparse map where the number of
items in each bin corresponds directly to the size of the bin, and the
maximum angular dispersion for each bin could be measured statistically
(by finding the maximum distance from each bin member to the bins
"origin" (where poses get truncated to). another measure would be the
maximumally spaced members in each bin. I'm worried that in some
cases some of the bins are too small or too large. Also would like to
investigate more the need to round instead of simply floor (I'm having
to add .5 to the rotation values before taking floor).