costmap_3d: query: fix fast path false negatives
The fast path was being taken when the new distance was over the
threshold. However, this is incorrect when the current cache distance
is below the threshold. Some other triangle may be in collision in
this case, so we must not take the fast path.

This issue was introduced by:
6067229b costmap_3d: query: improve fast-path performance